@@ -22,25 +22,23 @@
         "prepublishOnly": "npm run compile"
     },
     "dependencies": {
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^17.0.11",
+        "@types/react": "^18.0.0",
+        "@types/react-dom": "^18.0.0",
         "core-js": "^3.21.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
         "react-virtualized-auto-sizer": "^1.0.6",
         "react-window": "^1.8.6",
         "whatwg-fetch": "^3.6.2"
     },
     "devDependencies": {
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^17.0.11",
+        "@types/react": "^18.0.0",
+        "@types/react-dom": "^18.0.0",
         "@vitejs/plugin-react-refresh": "^1.3.6",
-        "pug": "^3.0.2",
         "sass": "^1.49.8",
         "ts-loader": "^9.2.6",
         "typescript": "^4.5.5",
         "vite": "^2.8.4",
-        "vite-plugin-pug": "^0.3.0",
         "webpack": "^5.69.1",
         "webpack-cli": "^4.9.2"
     }
