@@ -1,9 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from '../demo/component/App';
 
-ReactDOM.render(
-    <React.StrictMode>
-        <App />
-    </React.StrictMode>,
-document.getElementById('root'));
+const container = document.getElementById('app');
+const root = createRoot(container);
+root.render(<App />);
